--- pkg/apis/queue/v1alpha1/generated.proto
@@ -1,60 +1,0 @@
-
-// This file was autogenerated by go-to-protobuf. Do not edit it manually!
-
-syntax = 'proto2';
-
-package github.com.kube_queue.kube_queue.pkg.apis.queue.v1alpha1;
-
-import "github.com/gogo/protobuf/gogoproto/gogo.proto";
-import "k8s.io/api/core/v1/generated.proto";
-import "k8s.io/apimachinery/pkg/apis/meta/v1/generated.proto";
-import "k8s.io/apimachinery/pkg/api/resource/generated.proto";
-
-// Package-wide variables from generator "generated".
-option (gogoproto.marshaler_all) = true;
-option (gogoproto.stable_marshaler_all) = true;
-option (gogoproto.sizer_all) = true;
-option (gogoproto.goproto_stringer_all) = false;
-option (gogoproto.stringer_all) = true;
-option (gogoproto.unmarshaler_all) = true;
-option (gogoproto.goproto_unrecognized_all) = false;
-option (gogoproto.goproto_enum_prefix_all) = false;
-option (gogoproto.goproto_getters_all) = false;
-option go_package = "v1alpha1";
-
-message QueueUnit {
-  optional k8s.io.apimachinery.pkg.apis.meta.v1.ObjectMeta metadata = 1 [(gogoproto.customname) = "ObjectMeta", (gogoproto.nullable) = false];
-
-  optional Spec spec = 2 [(gogoproto.customname) = "Spec", (gogoproto.nullable) = false];
-
-  optional Status status = 3 [(gogoproto.customname) = "Status", (gogoproto.nullable) = false];
-}
-
-message QueueUnitList {
-  optional k8s.io.apimachinery.pkg.apis.meta.v1.ListMeta metadata = 1 [(gogoproto.customname) = "ListMeta", (gogoproto.nullable) = false];
-
-  repeated QueueUnit items = 2 [(gogoproto.customname) = "Items", (gogoproto.nullable) = false];
-}
-
-message Spec {
-  optional string type = 1 [(gogoproto.customname) = "JobType", (gogoproto.nullable) = false];
-
-  optional int32 priority = 2 [(gogoproto.customname) = "Priority", (gogoproto.nullable) = false];
-
-  optional string queue = 3 [(gogoproto.customname) = "Queue", (gogoproto.nullable) = false];
-
-  map<string, k8s.io.apimachinery.pkg.api.resource.Quantity> resource = 4 [(gogoproto.castkey) = "k8s.io/api/core/v1.ResourceName", (gogoproto.casttype) = "k8s.io/api/core/v1.ResourceList", (gogoproto.customname) = "Resource", (gogoproto.nullable) = false];
-}
-
-message Status {
-  optional string phase = 1 [(gogoproto.casttype) = "JobPhase", (gogoproto.customname) = "Phase", (gogoproto.nullable) = false];
-
-  optional string message = 2 [(gogoproto.customname) = "Message", (gogoproto.nullable) = false];
-}
-
-message GroupVersion {
-  optional string group = 1 [(gogoproto.customname) = "Group", (gogoproto.nullable) = false];
-
-  optional string version = 2 [(gogoproto.customname) = "Version", (gogoproto.nullable) = false];
-}
-
